Rename Location file to DataTypes, it contains now multiple data types which are special for JDWP.
Also rename MethodsReply to Methods to introduce more consistency with
the naming of other structs.
