--- OODebug/src/debugger/jdwp/Location.cpp
@@ -1,54 +1,0 @@
-/***********************************************************************************************************************
-**
-** Copyright (c) 2011, 2014 ETH Zurich
-** All rights reserved.
-**
-** Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
-** following conditions are met:
-**
-**    * Redistributions of source code must retain the above copyright notice, this list of conditions and the following
-**      disclaimer.
-**    * Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
-**      following disclaimer in the documentation and/or other materials provided with the distribution.
-**    * Neither the name of the ETH Zurich nor the names of its contributors may be used to endorse or promote products
-**      derived from this software without specific prior written permission.
-**
-**
-** THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
-** INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-** DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-** SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-** SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-** WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-** OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-**
-***********************************************************************************************************************/
-
-#include "Location.h"
-
-namespace OODebug {
-
-Location::Location(Protocol::TypeTagKind typeTag, qint64 classId, qint64 methodId, qint64 methodIndex)
-{
-	this->typeTag = typeTag;
-	this->classId = classId;
-	this->methodId = methodId;
-	this->methodIndex = methodIndex;
-}
-
-Location::~Location() {}
-
-bool Location::operator==(const Location &other) const
-{
-	return typeTag() == other.typeTag() && classId() == other.classId()
-			&& methodId() == other.methodId() && methodIndex() == other.methodIndex();
-}
-
-uint qHash(const Location &location) {
-	QByteArray data;
-	QDataStream stream(&data, QIODevice::ReadWrite);
-	stream << location;
-	return qHash(QString(data));
-}
-
-} /* namespace OODebug */
